backend-ml/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.0
redis==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
